[Test] Add test case base for unit and functional tests.
